requirements.txt
streamlit==1.32.2
pypdf==3.17.1
python-docx==0.8.11
google-generativeai==0.5.2
numpy==1.26.4
scikit-learn==1.3.2

